requirements.txt
setuptools~=57.0.0
matrix-synapse~=1.49.0